+++ src/rules_database.rs
@@ -2,6 +2,4 @@
 //!
 //! This module re-exports types from [`kingfisher_rules::rules_database`].
 
-pub use kingfisher_rules::rules_database::{
-    RuleDetectionProfileKind, RuleMatchProfile, RulesDatabase, format_regex_pattern,
-};
+pub use kingfisher_rules::rules_database::{RulesDatabase, format_regex_pattern};